Receive outer checksum info in the packet push
This changeset makes use of a new feature in flowenv allowing us to
receive headers along with the initial metadata from a specified
location in the packet.  For now we are just receiving the checksum
information from the MAC prepend.  This patch marks the L3 RX checksum
information in the packet metadata rather than passing it along to some
future processing and consuming a register.  It improves
receive-from-wire performance by about 800,000 pps.

I have looked at receiving packet headers as well in this initial push,
but we can only receive them at byte-level granularity and our
header-extract seems to require nicely aligned L3/L4 headers in order to
not run out of registers.