tests: redirect output of tcpdump
It seems like tcpdump has issues with loosing stderr, redirect
errors to a temporary file.  I was initially redirecting to
/dev/null but Jiayin provided a nice improvement to send it
to the temp dir along with all the other outputs.  Now if a
test fails one can take a look at the tcpdump_stderr.txt and
see if there are any packets received but not captured.

This siginificantly improves dependability of the tests for me.

Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>