test: use mv_to() to avoid problems running from localhost
When tests are run on localhost (HostA is localhost) we
can't copy files to /tmp using the same name and then
blindly remove the file.  In case of localhost we end up
with:

Which will either fail because cp will return 1 or even
if that didn't happen rm will remove the file and we will
be left without any copy of it.  We need to use proper
host-aware move (mv_to).

Note: mv_to() was added in NTI in a very recent revision!

Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>