Run the boot option from the autorun script
Currently, the nfp driver gets unloaded and loaded again during
system initialization. This is problematic since:

1) When Agilio OVS is installed on the system and loads during boot,
   this autorun process unloads the driver and restores CoreNIC again.
2) Third party network configuration tools might not expect netdevs
   to come up, and then be removed again, only to come up again a
   bit later - recent experiences with Network Manager indicates that
   this might be an issue with certain Network Manager configurations.
3) This seems like non-standard behaviour.

I can see at least one reason why there needs to be something
running after boot. The media configuration could have changed, and
in that case we want to make sure the udev rules are appropriately
updated. Ideally, the udev rules along with depmod config should only
happen once during installation, but at the moment the hwinfo doesn't
return all the MAC addresses available on a system when its configured
as a single port only. A future extension would be to improve the
udev rules to not rely on the MAC addresses at all.

This becomes a bit of a problem since the media can these days be
changed without a reboot - however, that probably needs to become
a CoreNIC documentation task then to request the user to rerun
the ns-core-nic-setup script.

So to address the 3 concerns listed above, run a specific procedure
after boot that only performs the necessary steps without
unnecessarily unloading the driver. The driver is just unloaded
if userspace CPP access is not available.

Signed-off-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>
Acked-by: Pablo Cascón <pablo.cascon@netronome.com>