Update flash version check
The latest BSP provides a different mechanism to check the BSP flash
version. The ns-core-nic-setup script is now used for this and called
from both Deb and RPM postinst scripts.

The flash version check has some minimum supported versions hardcoded.
If the minimum version is not met, recommend a full reflash. If only
the NSP minimum version is not met, or it is not the exact
NSP version supplied with the installed BSP (if there is a BSP
installed), then recommend to update the NSP only.

This resolves://NIC-236

Signed-off-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>
Acked-by: Pablo Cascón <pablo.cascon@netronome.com>