Static udev rules
* Instead of dynamically creating udev rules based on the currently used
MAC addresses, create a single fixed rule file that always allocates
nfp_pX netdev names sequentially from the first MAC address allocated
to the card.

This patch assumes that there will always be up to 8 MAC addresses
available on the card. This assumption is not valid for cards such as
Lithium/Carbon. But there is no harm in having the additional unused
rules always present. It is *very* unlikely that there will be more
than one card installed with overlapping MAC addresses. And in any case,
the naming scheme will need to change in the near term for multicard
cases in any case - the current scheme does not support multicard.

* Remove 'KERNEL=eth*' udev criteria

The addition of the KERNEL=eth* match criteria on the udev rules doesn't
seem to have any effect. However, on some systems (observed on some CentOS
7.3 machines), the udev rule is only matched when that is removed.

There have also been some statements on the web that Ubuntu 16.04 also
needs that particular match parameter removed.

comment://NFPBSP-2739

This was tested on CentOS 7.3, Ubuntu 14.04/16.04 with various kernel
versions.

* Removed the $attr{phys_port_name} from the udev rules for now. This
works well, but with breakout modes, the ports are named nfp_pXsY which
is inconsistent with the Agilio OVS netdev naming scheme and the CoreNIC
naming scheme on older kernels. So for the imminent AOVS 2.6.B release
(which needs to be highly coupled with CoreNIC) drop that new feature.

Signed-off-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>
Acked-by: Pablo Cascón <pablo.cascon@netronome.com>