systemd: Remove ExecStop implementation
The ExecStop gets executed whenever the ns-core-nic service gets
stopped, which happens on system reboot. This will clear the udev rules
prematurely.

So instead, remove the ExecStop completely. When removing the deb/RPM
package on a systemd distro, explicitly execute the ns-core-nic-setup
clean command.

I can't find a proper systemd hook that will get executed when the
service is disabled.

This resolves://NIC-252

Signed-off-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>
Acked-by: Pablo Cascón <pablo.cascon@netronome.com>