install: Compile error on one kernel makes the entire package installation fail
resolves://NIC-257

Modified the "deb_postint" DKMS template to NOT exit abruptly when some DKMS builds fail.

On some systems multiple kernels may be installed, but not all requirements are met when building DKMS modules against these kernels.
This update allows to failures, but displays a warning message that some builds for some kernels failed.

Signed-off-by: Pieter Gildenhuys <pieter.gildenhuys@netronome.com>
Acked-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>