Check nfp module ref count before trying to reload
When CoreNIC is installed as part of AOVS, and the CoreNIC package
is updated, ns-core-nic-setup will be executed again. However, in such a
case AOVS will have other kernel modules dependent on the nfp.ko module.
In which case it can't be unloaded and fails the setup procedure.

This could lead to the non-existance of udev rules.

However, when AOVS is running, CPP access will be available. So instead
of failing, check the ref count on the module and only fail if we don't
have CPP access with a non-zero ref count.

Similarly, we don't try to reload the module if the ref count is
non-zero. This will not succeed in any case.

Signed-off-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>
Acked-by: Pablo Cascón <pablo.cascon@netronome.com>