install: hide stderr during dkms remove
The following text is always printed when we install CoreNIC since we
always remove the driver from dkms irrespective of whether it has been
installed or not.

Error! There are no instances of module: nfp
2.0 located in the DKMS tree.

This has been observed to raise eyebrows from customers, see SDN-2207.

So instead, we hide any stderr output from the 'dkms remove' command.
The flipside of this is we lose some info during installation and
potentially info for debugging installation issues, but we haven't had
any issues on this front for some time. And the install part of the
installation is more important and likely to list valid issues.

wip://SDN-2207

Signed-off-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>
Acked-by: Pablo Cascón <pablo.cascon@netronome.com>