[app_master] fix regression in link state propagation
The merge of edelweiss into master branch (commit 5bbf7a2)
introduced a bug because the new link state handling code
in the master branch was not control VNIC ware. The logic
in lsc_send() was not updated to take this edelweiss
change into account. This patch also renames nic_intf to
port in lsc_check() for consistency.

Signed-off-by: Edwin Peer <edwin.peer@netronome.com>