[build] prevent clashes with upstream firmware
Install directory is renamed to prevent clashes with
upstream firmware package names. The installer also
takes care to save any existing firmware in the root
of /lib/firmware/netronome and restores the state on
unintall. The current strategy is not perfect. For
example, it will not properly handle multiple installs
of different agilio packages (each will save whatever
is there, clobbering the first one).

Signed-off-by: Edwin Peer <edwin.peer@netronome.com>