[bpf] update br_addr[] for latest devel SDK (build 3417)
The development branch of the SDK (6.x-devel) provided an
experimental and incompatible version of the br_addr[]
instruction in several builds (around the time frame of
build 3206). This code update is required to enable the
project to move forward to later SDK builds in order to
exploit other new features.

Later versions of the SDK changed to provide a mechanism for
automatic calculation of register liveness through code that
is not built/controlled by nfas. This liveness tracking extends
into call graph nodes designated by the visits[] token until
execution utlimately returns to SDK controlled code via the
rtn[] node. Further incompatible changes to this instruction
are still envisaged before it is finally declared stable, see
wip://THSDK-3557 for details.

Signed-off-by: Edwin Peer <edwin.peer@netronome.com>