[fix] mask BLS before set in packet from NFD path
The DATALEN field in NFD descriptors is 16 bits while packet length
is 14 bits. Normally, packet lengths from NFD are constrained to
fit and this mismatch could be ignored. But, for descriptors marked
as invalid due to LSO max segments being exceeded, the larger field
is copied directly from the host descriptor to aid debugging. This
can overflow into the neighboring BLS field in the application
packet metadata and so the BLS field is now masked to zero before
being set.

resolves://NIC-391

Signed-off-by: Edwin Peer <edwin.peer@netronome.com>