carbon: add support for 1x10G + 1x25G mode
The Carbon platform can support mixed mode operations, even though this
isn't generally an expected use case. However, we do want to support
this
in case users only change the media mode on a single port and reload the
driver, in which case, they would be unable to recover without manually
intervening.

The mixed mode configuration is essentially the same as the 2x25G
configuration, i.e.
- ME-FC is enabled;
- TM config is the same as the 2x25G mode.

This resolves://NIC-388.

Signed-off-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>
Acked-by: Edwin Peer <edwin.peer@netronome.com>