udev: temporary CentOS 7.3 udev fix
The current udev script and rules rely on the systemd provided rule
80-net-setup-link.rules to name the device. However, it was found that
the builtin systemd function, net_set_link, bails early due to a failure
to match any devices. This causes our netdevs to remain unnamed and
defaults to ethX names.

We still need to understand why this is happening. However, a crude fix
is to detect CentOS 7.3 installations and manually set the required
environment variable to the current netdev name. This is by no means a
proper solution, but considering this fix will not break any other
systems and resolve the issue currently seen on customer evaluation
systems, this is good enough for now.

This is a wip://NIC-406

Note, also fixed two minor scripting issues seen on CentOS 7.3.
If phys_port_name doesn't exist, the nfp-name-gen script would error out
early. And if UDEV_BIOSDEVNAME doesn't exist, the script would also
error out early.

Signed-off-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>
Acked-by: Edwin Peer <edwin.peer@netronome.com>