udev: remove CentOS 7.3 workaround
Remove the workaround added for NIC-406. Refer to the analysis in the
Jira ticket for more comprehensive information about the scenario.

But the script works as expected, consistent netdev naming was switched
off in the test system due to the presence of 'net.ifnames=0' in the
kernel commandline.

This resolves://NIC-406.

Signed-off-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>