[debug] add initial spec for ethtool dumps via kernel
The spec file (and the entirety of its generator script) borrows
heavily from the Disa project. The present solution is untidy
and presents a long term burden of refining and maintaining the
input spec. This functionality should ultimately move into SDK
tooling using linker intelligence and code annotations for
symbols.

resolves://BPF-80
resolves://NIC-387

Signed-off-by: Edwin Peer <edwin.peer@netronome.com>