[config] wait more agressively on MAC state changes
The longer waits serve as belts and brances for queues to drain
before potential soft resets. It is also a good idea to allow
per port configuration to stabilize before opening the taps on
the MAC, lest we unexpectedly drop packets after link up.

The more aggresive stalls cannot be afforded (nor should they
be needed) for config changes that do not affect MAC state.
Thus, the existing wait that was incurred for all PF config
messages has been removed.

- wip://NIC-432
- wip://NIC-438
- wip://NIC-441
- wip://NIC-442

Signed-off-by: Edwin Peer <edwin.peer@netronome.com>