[starfighter] specify an AMDA rather than full Harrier
For -chip nfp-6xxx-B0, the linker places symbols in memory that is not
present on the Starfigher platform. Specifying an AMDA number is more
precise and is suitable when targeting a specific hardware platform.

Note: the ":0" in the -chip option specifies which NFP in an AMDA that
might include more than one NFP in the future.  For now, :0 works for
existing platforms.

- resolves://NIC-454

Signed-off-by: Michael Rapson <michael.rapson@netronome.com>
Signed-off-by: Edwin Peer <edwin.peer@netronome.com>