[fix] ensure tx flush is enabled when link is down
Refactor the link state check code to separate VNIC state from
MAC link state. MAC RX state was improperly conflated with TX
flush. Thus, whenever a VNIC was enabled by the host it would
override the TX flush state maintained by the link monitoring
loop. This made it possible for all the system packet buffers
to get stuck in egress the TM queues of a port with a downed
link, starving the rest of the system of buffers.

- wip://NIC-438

Signed-off-by: Edwin Peer <edwin.peer@netronome.com>