[build] suppress arch dependent binaries warning
The firmware ELF files target the NFP and so it is appropriate
to produce an architecture independent RPM for the host operating
system. Normally doing this would produce an rpmbuild error,
which is overridden by the:

    %define _binaries_in_noarch_packages_terminate_build 0

directive. A warning, however, was still produced. This, in turn,
was upsetting the buildbot.

Signed-off-by: Edwin Peer <edwin.peer@netronome.com>