[app_master] wait for TMQ drain before TX flush
If the traffic manager TX queue is flushed while packets
are being transmitted, then it is possible to trasmit a
partial packet that is detected as an RX alignment error
by the remote party.

This patch attempts to address this issue by querying the
TMQ status and waiting for the queue to drain before
enabling TX flush.

wip://NIC-426

Signed-off-by: Edwin Peer <edwin.peer@netronome.com>