[pv] add support for pv_seek(..., EXIT_LABEL)
Parameterizing the pv_seek() macro with an exit label allows the
internal ctx_arb[]s to be exploited to efficiently create a hole
within the macro it is called from. This is not always more
efficient than manually creating the hole with an external branch
and depends on what the typical path through the seek macro would
be in the particular context of the call. In the case of the
T_INDEX_ONLY flag, using an exit label is more efficient if the
calling context doesn't have any instructions to replace the
required nops (the nops end up in the branch defer[] slots).

Signed-off-by: Edwin Peer <edwin.peer@netronome.com>
Reviewed-by: Ciaran Toal <ciaran.toal@netronome.com>