[actions] refactor MTU action into RX actions
Having individual RX actions for from host and from wire makes
the RX path symmetrical with the TX path actions. An additional
benefit (to be exploited later) is the ability to parameterize
RX actions. It is also marginally more efficient since the MTU
check for TX is now effectively pipelined.

Signed-off-by: Edwin Peer <edwin.peer@netronome.com>
Reviewed-by: Ciaran Toal <ciaran.toal@netronome.com>