[actions] remove the RX_CSUM action
Perform RX_CSUM unconditionally during the RX_WIRE action. The NBI
initialization code no longer maintains the MAC parser metadata that
this action utilizes. The minimal cost of computing these checksum
bits for the NFD descriptor make the tradeoff worthwhile, especially
when considering that the packet may be destined for a VF that does
not support CHECKSUM_COMPLETE.

Signed-off-by: Edwin Peer <edwin.peer@netronome.com>
Reviewed-by: Ciaran Toal <ciaran.toal@netronome.com>