[pv] refactor metadata to maintain offset stack
This patch maintains a stack of header offsets in preparation
for RSS over tunnel inners. The packet type if more efficiently
encodeing into a PROTOCOL field in order to make space for the
additional offset bits and the old MAC parse and Catamaran data
are no longer maintained. The protocol encoding is chosen in
such a way as to make RSS processing more efficient.

Signed-off-by: Edwin Peer <edwin.peer@netronome.com>
Reviewed-by: Ciaran Toal <ciaran.toal@netronome.com>