[buildbot] Add bpf-test builder with custom mail notifier
Provide the driver test steps as generic steps to be included in some
other builders. At the moment, the BPF test and general driver tests are
the same, but they will differ in the near future.

BPF tests will now be executed nightly at 6pm. A specific mail notifier
has been configured for this.

Signed-off-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>
Acked-by: Edwin Peer <edwin.peer@netronome.com>