[app_master] support RSS parameterization
The RSS action recently gained support for variable sized RSS tables
stored in NN registers, allowing the mapping tables for more ports
to fit. This patch adds support for 4 bit and 2 bit table entries
reducing the space required per port to 16 words and 8 words
respectively when the number of queues is limited (as they are when
more ports are configured).

resolves://NIC-477
resolves://NIC-532

Signed-off-by: Edwin Peer <edwin.peer@netronome.com>