[config] add VEB table entry control message marshaller
The VEB table reuses the BPF map code, which does not presently
have a MicroC implementation. BPF map entries are manipulated
via an auxiliary ME written in microcode that handles control
messages from the host.

For SR-IOV VEB entries, a control message to this auxiliary ME
is generated in order to add and remove table entries in response
to control BAR updates in the app_master MicroC context. This
patch comprises the plumbing needed to marshal these messages
to the control message handler ME.

wip://NIC-500
resolves://NIC-516

Signed-off-by: Norm Bagley <norman.bagley@netronome.com>
Reviewed-by: Edwin Peer <edwin.peer@netronome.com>