[config] add VEB action sequences for VFs
The app_config_sriov_port() function builds a list of actions that
will be installed in the VEB table (to be performed when delivering
packets to a VF). The VEB table will also be used to direct packets
from VFs to the PF and so the portions of the action list that are
common have been refactored into app_config_pf_common().

wip://NIC-500

Signed-off-by: Norm Bagley <norman.bagley@netronome.com>
Signed-off-by: Ciaran Toal <ciaran.toal@netronome.com>
Reviewed-by: Edwin Peer <edwin.peer@netronome.com>