[config] add handle_sriov_update() function
The handle_sriov_update() function configures the VEB table in
accordance with the VF's MAC and VLAN configuration supplied
via the PF ABI.

This function will be called in response to config updates in
the VF BAR. There are some open TODO items that still need to
be resolved in future commits. Updates are also required to
insert all VLAN combinations when matching only MAC (without
VLAN) in order to support wip://NIC-491.

wip://NIC-500

Signed-off-by: Norm Bagley <norman.bagley@netronome.com>
Signed-off-by: Ciaran Toal <ciaran.toal@netronome.com>
Reviewed-by: Edwin Peer <edwin.peer@netronome.com>