[app_master] add VF link state handling
Link state for VFs can be configured in 3 modes. Each VF can be
configured to be always up, down or to follow the link state of
the associated physical port. In the latter case, the VFs that
are associated with a given physical port are maintained in the
multidimensional vf_lsc_list array which indicates the VFs that
must be updated on link state changes.

Signed-off-by: Norm Bagley <norman.bagley@netronome.com>
Signed-off-by: Ciaran Toal <ciaran.toal@netronome.com>
Reviewed-by: Edwin Peer <edwin.peer@netronome.com>