[config] add multiple entries to VEB table for mac only matching
The VEB table lookup for received packets does a lookup using the
MAC and VLAN tag. When no VLAN is configured for the VF, then we
need to match just the MAC. This patch adds VEB table entries for
all possible 12 bit VLAN tags when no VLAN is configured, so a
match will occur with the configured MAC and any VLAN tag.

resolves://NIC-491
wip://NIC-500

Signed-off-by: Norm Bagley <norman.bagley@netronome.com>
Reviewed-by: Edwin Peer <edwin.peer@netronome.com>