[fix] add RX_WIRE / RX_HOST instruction before DROP
pv_init_nbi() and pv_init_nfd() are now called by RX_WIRE and
RX_HOST actions so that their behavior can be parameterized by
the control plane. The consequence of this design change is
that explicit RX_WIRE and RX_HOST actions (as appropriate) are
always necessary to initialize packet metadata that is required
to DROP a packet.

Care must be taken to ensure the action list is appropriately
initialized before packets hit the action interpreter as well
as whenever queues are reconfigured.

wip://BPF-178
resolves://NIC-568
resolves://NIC-573

Signed-off-by: Mary Pham <mary.pham@netronome.com>
Reviewed-by: Edwin Peer <edwin.peer@netronome.com>