[nfd] support send without freeing CTM buffer
This out-of-tree patch treats BLQ 3 as special multicast buffer
list. Packets released to this queue are reference counted by
the application, signalling to NFD that it should not free the
CTM buffer after sending the packet.

For the purposes of CoreNIC, the multicast BLQ is hardcoded to 3.
This approach is obviously not suitable for upstream NFD. At very
least, this special behavior will need to be made optional. Using
a magic BLQ number in this fashion was expedient, however, it may
also be more appropriate to devise a more generic signalling
mechanism in future.

wip://NIC-555

Signed-off-by: Edwin Peer <edwin.peer@netronome.com>
Reviewed-by: Ciaran Toal <ciaran.toal@netronome.com>