[pkt_io] implement non-terminal multicast TX actions
The TX macros are now parameterized with options allowing for
continued action processing after packet transmission. Multicast
is thus achieved by executing multiple TX macros in succession.
Such multicast packets are assinged to a special multicast aware
BLQ where packet buffers are tracked and reference counted
each time the packet is re-transmitted.

Two modes for continued processing are supported. In the
simplest case, action processing unconditionally continues
after transmission of the packet. The second mode supports
simple logic that only continues action processing if the
packet is destined for a group MAC.

The packet metadata was significantly refactored as part of
this patch laying the groundwork to ultimately support the
sending different versions of the packet to different
destinations. To that end, the packet metadata has been
compacted to 8 words. The model for prepend metadata has
also been cleaned up. Now prepend metadata is stored in the
packet vector until such time as the packet is dispatched
for transmission (permitting packet prepend modifications
to be performed without regard to the prepend metadata).

wip://NIC-555

Signed-off-by: Edwin Peer <edwin.peer@netronome.com>
Reviewed-by: Ciaran Toal <ciaran.toal@netronome.com>