[actions] add packet stack for multicast variations
When multicasting a packet to VLAN tagged vs untagged
endpoints, it becomes necessary to send slightly different
versions of the same packet to their respective destinations.

One way this could be accomplished is to wait until a given
version has been sent to all of its associated destinations
before modifying the packet and sending it along to the next
set of destinations. Such an approach introduces serial
latency and presents challenges if some of the destinations
are flow controlled or otherwise delayed. In general a model
where each variant of the packet is allocated its own logical
buffer is desired.

The packet push and pop actions keep track of packet
variants in a logical stack. A given version of the packet
can be sent to multiple destinations after which it can be
pushed to the stack before making further modificaitons for
subsequent destinations. The implementation exploits the
NFP's split buffer architecture to accomplish this without
actually allocating any new buffers.

This implementation choice, however, imposes some minor
limitations:

1) The maximum packet stack depth is limited to one. This
limited stack is sufficient for present needs and avoids
the complexities associated with allocating additional
packet buffers. The model does generalize and could
conceivably be extended without control plane changes to
accomodate N modifications at the cost of a more complex
implementation that manages multiple CTM buffers.

2) The NBI's reliance on a CTM buffer implies the further
limitation that packets being sent out the wire must be
the first variant of the packet (ie. before being pushed
or after being popped). Thus far it has been possible to
order the modifications and transmissions appropriately
such that this constraint is satisfied.

3) Furthermore, because the first variant resides in the
CTM buffer, the modifications to this first variant are
restricted to the part of the packet present in CTM. In
practice, useful modifications are invariably performed
at the head end of the packet.

resolves://NIC-555

Signed-off-by: Edwin Peer <edwin.peer@netronome.com>
Reviewed-by: Ciaran Toal <ciaran.toal@netronome.com>