[actions] add support for updating L3/L4 checksums
The NFP provides hardware checksum insertion only for packets
egressing the MAC block. Traffic sent between host destinations
does not traverse the MAC and thus a software implementation of
checksum insertion is required. This feature is also useful for
packet encapsulations not supported by the MAC.

The implementation leverages the existing checksum complete
logic to calculate L4 checksums. To construct an L4 sum, a
second pass over the head of the packet is performed. This
result is then deducted from the complete checksum covering
the whole packet. IPv4 checksums are supported in the obvious
fashion.

The checksum action is parameterized by a set of bits (one for
each type checksum supported) that mirror the checksum offload
bits set by host sources. The offload bits are used to mask the
action arguments so that checksum insertion only occurs for host
PCIe based sources.

resolves://NIC-578

Signed-off-by: Edwin Peer <edwin.peer@netronome.com>
Signed-off-by: Ciaran Toal <ciaran.toal@netronome.com>