[build] report name instead of fw_name in 'ethtool -i'
The MIP name reported to the driver should contain the fw_id
string. The recently added '-fw_name' argument which is used
to resolve firmware flash priorities overrides the MIP name
if it appears after '-name' in the nfld argument list. This
patch reverses the arguments in the build so that the correct
firmware version string is once again reported to the driver.

Signed-off-by: Edwin Peer <edwin.peer@netronome.com>