[app_master] refactor to reduce code store utilization
The nic_local_epoch() feature is not yet in active use, but
in any event, for its intended purpose of synchronizing the
datapath with config changes, it can safely be called from
one place instead of two. It is safe to trigger the
synchronization on NFP_NET_CFG_CTRL_ENABLE state changes.

Signed-off-by: Edwin Peer <edwin.peer@netronome.com>