[pv] never drop bad packets
Catamaran is kind enough to mark packets exhibiting various
errors, such as invalid IPv4 checksums or headers with
illegal lengths. Some of these packets result in an error
bit being set in the ingress packet metadata while others
are assigned to the unordered sequencer context.

In the interests of expedience and the fact that practical
users typically don't care for these packets anyway,
previous versions would simply drop such packets. A good
NIC, however, should not discriminate as a matter of
policy and should rather forward any badly formed packets
faithfully onwards to the host.

In the interests of simplicity of the egress code path,
the data path was designed to never skip the global reoder
block, thus presenting a challenge for sending packets that
have no associated sequence number. This is resolved here
by assigning a fake sequence number derived from an atomic
increment of a global counter. The allocation of additional
sequence contexts was also required to accomodate this new
modality.

wip://NIC-582

Signed-off-by: Edwin Peer <edwin.peer@netronome.com>