[pv] fix mask in UDP tunnel check
A header parse should be executed if the packet is UDP and VXLAN
or GENEVE tunnels have been enabled. The configuration options
in the low bits are subtracted from zero to create an all ones
mask covering the bits in the word where Catamaran stores the
L4 classification information about the packet. This mask was
not being saved and therefore classification bits were always
being masked to zero with the result that the required header
parse would never run.

Signed-off-by: Edwin Peer <edwin.peer@netronome.com>