[actions] use preprocessor to avoid __actions_t_idx updates
During action processing, care must be taken to restore T_INDEX
each time a context swaps out. The __actions_t_idx register is
utilized as a cursor to track the next action to be processed.
It is advanced for each __actions_read() call with a call to
__actions_t_idx_restore() updating the T_INDEX CSR to reflect
all the reads since the time that the action list was loaded.

Preprocessor accounting can be used in cases where multiple
__action_read() occur in succession in order to avoid unecessary
updates to the  cursor. This can now be achieved by simply
grouping multiple calls to __action_read() between a calls to
__actions_read_begin() and __actions_read_end().

Signed-off-by: Edwin Peer <edwin.peer@netronome.com>