[build] share code memory between app_master and cmsg handler
The app_master is tight on code store and needs more room to
accomodate upcomming changes. The control message handler for
BPF applications is presently using less than half of its
codestore and is not particularly performance sensitive (it
can afford a small branching penalty). This patch reshuffles
the build layout in order to place app_master and the mapcmsg
handler on neighboring MEs and enables linker codeshare for
their deployment.

Signed-off-by: Edwin Peer <edwin.peer@netronome.com>