[actions] switch order of checksum and rss
The app master was recently changed to perform checksum
before BPF and RSS actions to more closely match the
behavior of host side XDP. The order was initially the
other way around because the checksum metadata could not
be directly witten to the front of the packet (due to
the possibility of BPF adjusting the head offset).

Having the actions in code store order is more optimal
for the interpreter and thus the code layout should
most closely reflect the typical case.

Signed-off-by: Edwin Peer <edwin.peer@netronome.com>