[udev] exit if interfaces are already named
When the agilio-naming-policy package is installed on a
RHEL/CentOS system that already renames the NFP interfaces
using the native phys-port-name-gen script, the environment
variables used for naming will already have the phys_port_name
attributed appended. We check for this in a dash compatible
way and exit silently if it's already present.

resolves://NIC-577.

Signed-off-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>
Reviewed-by: Edwin Peer <edwin.peer@netronome.com>