[nfd] correctly configure work queue size
The configured size of the NFD work queue is derived
from the size of the buffer list queues that can feed
into it. The work queue must be sufficiently large to
fit all the associated buffers in the system.

The MU buffers for NFD were isolated from those used by
the NBI in commit ee17b9601, however, the function that
defines NFD_IN_WQ_SZ was not updated accordingly. This
bug did not manifest immediately since BLQ1 happened to
be the same size as the original BLQ. Later, in commit
0b1a0c439, the size of BLQ1 was increased to no longer
match that of BLQ0. This meant that it was possible for
the work queue to overflow, losing packets and locking
up the transmit queue due to missing sequence numbers.

resolves://NIC-634
resolves://NIC-644
resolves://NIC-652
resolves://NIC-645

Signed-off-by: Edwin Peer <edwin.peer@netronome.com>