[build] generate flavor specific initramfs hooks
The Debian packages for different build flavors each include
the same initramfs hook script. This would result in a failure
to install the package, due to the file name conflict, if a
another flavor of the firmware was previously installed.

To resolve the naming conflict, this patch gives the initramfs
hook a unique name in each flavor's package. Note, although it
is possible to install multiple flavors simultaneously on a
given machine, the most recently installed package will take
precedence by updating the symlinks in /lib/firmware. Thus the
driver will load the most recently installed firmware. These
symlinks need to be maintained by the user in order to select
between different firmware loads.

resolves://NIC-649

Signed-off-by: Edwin Peer <edwin.peer@netronome.com>