[app_master] wait for link up on PF VNIC enable
The kernel driver enables link via an NSP call immediately prior to
bringing up a PF VNIC. By waiting for the link to come up before
signalling the VNIC enabled event to the driver, the dropping of the
first packets in the host's transmit queue can be avoided. This wait
is best effort - the implementation will give up after 2 seconds if
the link does not come up. It would be preferable to disable NFD
queues while the link is down in future so that descriptors do not
get DMAed from the host in this state.

This patch includes a Flowenv fix to the sleep() API. Sleeps were
shorter than expected, possibly causing other races with the NSP.

wip://NIC-575

Signed-off-by: Edwin Peer <edwin.peer@netronome.com>