[app_master] write whole VXLAN port table to NNRs
NFP_NET_CFG_VXLAN_SZ is defined in bytes over 2 byte table entries.
Instead, make use of NFP_NET_N_VXLAN_PORTS to determine the number
of ports. The writing of zeroes to unused entries is not required
for correctness, but aids in debugging and is justifiable for
config plane code.

Signed-off-by: Edwin Peer <edwin.peer@netronome.com>