[tuning] improve buffer configuration
This new buffer config results in significantly improved performance.
In the RX direction, all buffers are now allocated from the same MU.
The rationale is that in this direction small packets will reside
entirely within CTM. In the other direction, a larger number of
distributed buffers are allocated. On a Dell R730 with Xeon E5-2630
CPUs this change results in an overall improvement of 3.6Mpps when
packets are looped back using XDP in the host driver. RX and TX
directions are now also better balanced under full-duplex traffic.

Signed-off-by: Edwin Peer <edwin.peer@netronome.com>