[actions] fix invalid offset test in L4 checksum
The parser can produce a valid IP offset while the L4 header is unknown
and zero. Simultaneously testing both offsets for zero is incorrect. It
should be ok to only test the L4 offset for zero, since the parser will
not generate a legal L4 offset without generating a valid IP offset.

This patch also fixes misleading comments in the same region.

Signed-off-by: Bill McGee <bill.mcgee@netronome.com>
Reviewed-by: Edwin Peer <edwin.peer@netronome.com>