[docs] import 2.0.7 RST User Guide
This commit contains the user guide files in rst format.

The Makefile in the /docs folder has been updated to build the document
via the 'user-guide' make target.

The conf.py setting of basic theme, and the build setting of singlehtml are
required for the built document to be uploaded to the freshdesk site in an
automated manner.

Note, the sphinxcontrib-imageembed pip package is required to build the
document, and sphinxcontrib-spelling is required to run make
user-guide-spelling.

Signed-off-by: Brendan Galloway <brendan.galloway@netronome.com>
Signed-off-by: Edwin Peer <edwin.peer@netronome.com>