stats: fix termination in VNIC stats aggregation
The statistics aggregation logic maintains per VNIC counters in
an MU backed region that is read and written using the bulk engine.
These backing stats are copied to and from local memory where
queue stats are aggregated serially for each VNIC. The logic is
written to support backing regions that are arbitrarily sized
relative to the size of the read/write XFER blocks. In practice,
the backing region size is rounded up to the next power of two
for efficient indexing.

The calculation of the remaining bytes for the last copied block
was off by one, copying one too few elements. This was discovered
by inspection while investigating wip://NIC-679. Fortunately, this
issue never manifested in any existing builds because the backing
region was not filled to capacity by statistics.

Signed-off-by: Edwin Peer <edwin.peer@netronome.com>